backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

SQLAlchemy>=2.0.35,<3
asyncpg==0.29.0
psycopg[binary]==3.2.9
alembic==1.13.1

redis==5.0.4
celery==5.3.6

qdrant-client>=1.15.1,<2
openai==1.37.0
tiktoken==0.7.0
httpx==0.27.0

structlog==24.1.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
prometheus-client==0.19.0

python-jose[cryptography]==3.3.0
PyJWT==2.8.0

web3==6.19.0
eth-account==0.10.0
eth-utils==4.1.0

supabase==2.4.0

python-docx==1.1.2
reportlab==4.2.0
pdfminer.six==20231228
pypdf==4.2.0
pytesseract==0.3.10
Pillow==10.3.0

pytest==8.2.0

